#13 refactor project
