+++ app/components/BottomMenu.js
