+++ package.json
@@ -4,8 +4,8 @@
   "main": "index.js",
   "scripts": {
     "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web"
   },
   "dependencies": {
@@ -23,6 +23,9 @@
     "crypto-js": "^4.2.0",
     "expo": "^53.0.4",
     "expo-app-loading": "^1.0.3",
+    "expo-auth-session": "^6.1.5",
+    "expo-clipboard": "~7.1.4",
+    "expo-dev-client": "~5.1.8",
     "expo-font": "~13.3.0",
     "expo-status-bar": "~2.2.3",
     "install": "^0.13.0",
@@ -42,8 +45,7 @@
     "react-native-url-polyfill": "^2.0.0",
     "react-native-vector-icons": "^10.2.0",
     "react-native-web": "^0.20.0",
-    "zustand": "^5.0.3",
-    "expo-clipboard": "~7.1.4"
+    "zustand": "^5.0.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
